pyproject.toml
[project]
name = "testcase-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "approvaltests>=15.0.0",
    "black>=25.11.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.17.0",
    "pre-commit>=4.4.0",
    "pytest>=9.0.2",
    "pytest-approvaltests>=0.2.4",
    "pyyaml>=6.0.3",
    "tcms-api>=15.0",
]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude=".git,__pycache__,build,dist,.venv"
